requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
click==8.0.4
    # via pip-tools
joblib==1.1.0
    # via scikit-learn
numpy==1.22.3
    # via
    #   pandas
    #   scikit-learn
    #   scipy
pandas==1.4.1
    # via -r requirements.in
pep517==0.12.0
    # via pip-tools
pip-tools==6.5.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
python-dotenv==0.20.0
    # via -r requirements.in
pytz==2022.1
    # via pandas
scikit-learn==1.0.2
    # via -r requirements.in
scipy==1.8.0
    # via scikit-learn
six==1.16.0
    # via python-dateutil
threadpoolctl==3.1.0
    # via scikit-learn
tomli==2.0.1
    # via pep517
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
